# evc-app/src/_assets/api.yml
swagger: "2.0"
info:
  description: AUA backend API
  version: "1.0.0"
  title: "AUA API"
basePath: "/api/v1"
schemes:
  - http
  - https
...
paths:
...
  /auth/reset_password/{token}:
    get:
      operationId: retrievePassword
      # security:
      #   - authAnyRole: []
      parameters:
        - name: token
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
...
  /recurring/:id:
    get:
      operationId: getRecurring
      # security:
      #   - authClient: []
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
    delete:
      operationId: deleteRecurring
      # security:
      #   - authClient: []
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
  /recurring/:id/run:
    post:
      operationId: runRecurring
      # security:
      #   - authClient: []
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
...
  /message/:id:
    get:
      operationId: getMessage
      # security:
      #   - authClient: []
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
...